requirements.txt
fastapi
uvicorn
openai
streamlit
tavily-python
python-dotenv
requests